+++ examples/random-generation-protocol/src/lib.rs
@@ -3,7 +3,7 @@
 #![no_std]
 #![forbid(unused_crate_dependencies, missing_docs)]
 
-#[cfg(feature = "std")]
+#[cfg(any(feature = "std", test))]
 extern crate std;
 
 extern crate alloc;
@@ -173,14 +173,16 @@ pub struct Blame {
 
 #[cfg(test)]
 mod tests {
-    use alloc::vec;
+    use alloc::{vec, vec::Vec};
 
+    use rand::Rng;
     use round_based::simulation::Simulation;
+    use sha2::{Digest, Sha256};
 
     use super::{protocol_of_random_generation, Msg};
 
     #[tokio::test]
-    async fn main() {
+    async fn simulation_async() {
         let mut rng = rand_dev::DevRng::new();
 
         let n: u16 = 5;
@@ -202,5 +204,150 @@ mod tests {
         }
 
         std::println!("Output randomness: {}", hex::encode(output[0]));
+    }
+
+    #[test]
+    fn simulation_sync() {
+        let mut rng = rand_dev::DevRng::new();
+
+        let simulation = round_based::simulation::SimulationSync::from_async_fn(5, |i, party| {
+            protocol_of_random_generation(party, i, 5, rng.fork())
+        });
+
+        let outputs = simulation
+            .run()
+            .unwrap()
+            .into_iter()
+            .collect::<Result<Vec<_>, _>>()
+            .unwrap();
+        for output_i in &outputs {
+            assert_eq!(*output_i, outputs[0]);
+        }
+    }
+
+    // Emulate the protocol using the state machine interface
+    #[test]
+    fn state_machine() {
+        use super::{CommitMsg, DecommitMsg, Msg};
+        use round_based::{
+            state_machine::{ProceedResult, StateMachine},
+            Incoming, Outgoing,
+        };
+
+        let mut rng = rand_dev::DevRng::new();
+
+        let party1_rng: [u8; 32] = rng.gen();
+        let party1_com = Sha256::digest(party1_rng);
+
+        let party2_rng: [u8; 32] = rng.gen();
+        let party2_com = Sha256::digest(party2_rng);
+
+        // Start the protocol
+        let mut party0 = round_based::state_machine::wrap_protocol(|party| async {
+            protocol_of_random_generation(party, 0, 3, rng).await
+        });
+
+        // Round 1
+
+        // Party sends its commitment
+        let ProceedResult::SendMsg(Outgoing {
+            msg: Msg::CommitMsg(party0_com),
+            ..
+        }) = party0.proceed()
+        else {
+            panic!("unexpected response")
+        };
+
+        // Round 2
+
+        // Party needs messages sent by other parties in round 1
+        let ProceedResult::NeedsOneMoreMessage = party0.proceed() else {
+            panic!("unexpected response")
+        };
+        // Provide message from party 1
+        party0
+            .received_msg(Incoming {
+                id: 0,
+                sender: 1,
+                msg_type: round_based::MessageType::Broadcast,
+                msg: Msg::CommitMsg(CommitMsg {
+                    commitment: party1_com,
+                }),
+            })
+            .unwrap();
+        let ProceedResult::NeedsOneMoreMessage = party0.proceed() else {
+            panic!("unexpected response")
+        };
+        // Provide message from party 2
+        party0
+            .received_msg(Incoming {
+                id: 1,
+                sender: 2,
+                msg_type: round_based::MessageType::Broadcast,
+                msg: Msg::CommitMsg(CommitMsg {
+                    commitment: party2_com,
+                }),
+            })
+            .unwrap();
+
+        // Party sends message in round 2
+        let ProceedResult::SendMsg(Outgoing {
+            msg: Msg::DecommitMsg(party0_rng),
+            ..
+        }) = party0.proceed()
+        else {
+            panic!("unexpected response")
+        };
+
+        {
+            // Check that commitment matches the revealed randomness
+            let expected = Sha256::digest(party0_rng.randomness);
+            assert_eq!(party0_com.commitment, expected);
+        }
+
+        // Final round
+
+        // Party needs messages sent by other parties in round 2
+        let ProceedResult::NeedsOneMoreMessage = party0.proceed() else {
+            panic!("unexpected response")
+        };
+        // Provide message from party 1
+        party0
+            .received_msg(Incoming {
+                id: 3,
+                sender: 1,
+                msg_type: round_based::MessageType::Broadcast,
+                msg: Msg::DecommitMsg(DecommitMsg {
+                    randomness: party1_rng,
+                }),
+            })
+            .unwrap();
+        let ProceedResult::NeedsOneMoreMessage = party0.proceed() else {
+            panic!("unexpected response")
+        };
+        // Provide message from party 2
+        party0
+            .received_msg(Incoming {
+                id: 3,
+                sender: 2,
+                msg_type: round_based::MessageType::Broadcast,
+                msg: Msg::DecommitMsg(DecommitMsg {
+                    randomness: party2_rng,
+                }),
+            })
+            .unwrap();
+        // Obtain the protocol result
+        let ProceedResult::Output(Ok(output_rng)) = party0.proceed() else {
+            panic!("unexpected response")
+        };
+
+        let output_expected = party0_rng
+            .randomness
+            .iter()
+            .zip(&party1_rng)
+            .zip(&party2_rng)
+            .map(|((a, b), c)| a ^ b ^ c)
+            .collect::<alloc::vec::Vec<_>>();
+        assert_eq!(output_rng, output_expected.as_slice());
     }
 }