+++ wasm/no_std/src/lib.rs
@@ -3,6 +3,7 @@
 use core::panic::PanicInfo;
 
 pub use random_generation_protocol::*;
+pub use round_based;
 
 #[panic_handler]
 fn panic(_panic: &PanicInfo<'_>) -> ! {